Add files via upload
Web scraper is a tool or software that extracts data from websites. It sends a request to a webpage, parses the HTML content, and retrieves specific information based on predefined criteria. Web scrapers can automate the process of gathering large amounts of data from the web efficiently. We have used BeautifulSoup which is a Python library used for parsing HTML and XML documents. The provided script uses 'requests' to fetch a webpage and BeautifulSoup to parse and extract news headlines. This combination allows for efficient and straightforward web scraping tasks.